requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.19.2
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
redis==4.2.2
requests==2.27.1
icecream==2.1.2